refactor(layout): extract grid to mixin and remove layout components
BREAKING CHANGE:
`sbb-grid`, `sbb-section` and `sbb-stack` were removed.
`sbb-stack` and `sbb-section` can easily be replaced by css flex or grid on consumer side.
As replacement of the `sbb-grid` component there is now a css class and sass mixin.
