+++ react-library/src/components/stencil-generated/index.ts
@@ -27,7 +27,6 @@ export const SbbDivider = /*@__PURE__*/createReactComponent<JSX.SbbDivider, HTML
 export const SbbFooter = /*@__PURE__*/createReactComponent<JSX.SbbFooter, HTMLSbbFooterElement>('sbb-footer');
 export const SbbFormError = /*@__PURE__*/createReactComponent<JSX.SbbFormError, HTMLSbbFormErrorElement>('sbb-form-error');
 export const SbbFormField = /*@__PURE__*/createReactComponent<JSX.SbbFormField, HTMLSbbFormFieldElement>('sbb-form-field');
-export const SbbGrid = /*@__PURE__*/createReactComponent<JSX.SbbGrid, HTMLSbbGridElement>('sbb-grid');
 export const SbbHeader = /*@__PURE__*/createReactComponent<JSX.SbbHeader, HTMLSbbHeaderElement>('sbb-header');
 export const SbbHeaderAction = /*@__PURE__*/createReactComponent<JSX.SbbHeaderAction, HTMLSbbHeaderActionElement>('sbb-header-action');
 export const SbbIcon = /*@__PURE__*/createReactComponent<JSX.SbbIcon, HTMLSbbIconElement>('sbb-icon');
@@ -51,11 +50,9 @@ export const SbbPearlChainVertical = /*@__PURE__*/createReactComponent<JSX.SbbPe
 export const SbbPearlChainVerticalItem = /*@__PURE__*/createReactComponent<JSX.SbbPearlChainVerticalItem, HTMLSbbPearlChainVerticalItemElement>('sbb-pearl-chain-vertical-item');
 export const SbbRadioButton = /*@__PURE__*/createReactComponent<JSX.SbbRadioButton, HTMLSbbRadioButtonElement>('sbb-radio-button');
 export const SbbRadioButtonGroup = /*@__PURE__*/createReactComponent<JSX.SbbRadioButtonGroup, HTMLSbbRadioButtonGroupElement>('sbb-radio-button-group');
-export const SbbSection = /*@__PURE__*/createReactComponent<JSX.SbbSection, HTMLSbbSectionElement>('sbb-section');
 export const SbbSector = /*@__PURE__*/createReactComponent<JSX.SbbSector, HTMLSbbSectorElement>('sbb-sector');
 export const SbbSignet = /*@__PURE__*/createReactComponent<JSX.SbbSignet, HTMLSbbSignetElement>('sbb-signet');
 export const SbbSlider = /*@__PURE__*/createReactComponent<JSX.SbbSlider, HTMLSbbSliderElement>('sbb-slider');
-export const SbbStack = /*@__PURE__*/createReactComponent<JSX.SbbStack, HTMLSbbStackElement>('sbb-stack');
 export const SbbTabAmount = /*@__PURE__*/createReactComponent<JSX.SbbTabAmount, HTMLSbbTabAmountElement>('sbb-tab-amount');
 export const SbbTabGroup = /*@__PURE__*/createReactComponent<JSX.SbbTabGroup, HTMLSbbTabGroupElement>('sbb-tab-group');
 export const SbbTabTitle = /*@__PURE__*/createReactComponent<JSX.SbbTabTitle, HTMLSbbTabTitleElement>('sbb-tab-title');
