--- src/components/sbb-grid/readme.md
@@ -1,16 +1,0 @@
-# sbb-grid
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property     | Attribute    | Description        | Type                                                               | Default     |
-| ------------ | ------------ | ------------------ | ------------------------------------------------------------------ | ----------- |
-| `appearance` | `appearance` | Section appearance | `"primary" \| "primary-negative"`                                  | `'primary'` |
-| `variant`    | `variant`    | Grid variant       | `"base" \| "base--eight-columns-centered" \| "base--top-products"` | `'base'`    |
-
-
-----------------------------------------------
-
-
