--- src/components/sbb-grid/sbb-grid.custom.d.ts
@@ -1,4 +1,0 @@
-export interface InterfaceGridAttributes {
-  appearance: 'primary' | 'primary-negative';
-  variant: 'base' | 'base--top-products' | 'base--eight-columns-centered';
-}
