--- src/components/sbb-grid/sbb-grid.e2e.ts
@@ -1,13 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-
-describe('sbb-grid', () => {
-  let element, page;
-
-  it('renders', async () => {
-    page = await newE2EPage();
-    await page.setContent('<sbb-grid></sbb-grid>');
-
-    element = await page.find('sbb-grid');
-    expect(element).toHaveClass('hydrated');
-  });
-});
