--- src/components/sbb-section/readme.md
@@ -1,16 +1,0 @@
-# sbb-section
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property     | Attribute    | Description        | Type                                                                                                     | Default                 |
-| ------------ | ------------ | ------------------ | -------------------------------------------------------------------------------------------------------- | ----------------------- |
-| `appearance` | `appearance` | Section appearance | `"primary" \| "primary-negative"`                                                                        | `'primary'`             |
-| `width`      | `width`      | Section width      | `"full-bleed--forever" \| "full-bleed--until-ultra" \| "full-bleed--until-ultra-plus" \| "page-spacing"` | `'full-bleed--forever'` |
-
-
-----------------------------------------------
-
-
