--- src/components/sbb-section/sbb-section.custom.d.ts
@@ -1,8 +1,0 @@
-export interface InterfaceSectionAttributes {
-  appearance: 'primary' | 'primary-negative';
-  width:
-    | 'full-bleed--forever'
-    | 'full-bleed--until-ultra-plus'
-    | 'full-bleed--until-ultra'
-    | 'page-spacing';
-}
