--- src/components/sbb-section/sbb-section.e2e.ts
@@ -1,13 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-
-describe('sbb-section', () => {
-  let element, page;
-
-  it('renders', async () => {
-    page = await newE2EPage();
-    await page.setContent('<sbb-section></sbb-section>');
-
-    element = await page.find('sbb-section');
-    expect(element).toHaveClass('hydrated');
-  });
-});
