--- src/components/sbb-stack/sbb-stack.chromatic.stories.js
@@ -1,6 +1,0 @@
-import config, * as stories from './sbb-stack.stories';
-import { chromaticStoryConfig, combineStories } from '../../global/helpers/chromatic';
-
-export const chromaticStories = combineStories(config, stories);
-
-export default chromaticStoryConfig(config);
