--- src/components/sbb-stack/sbb-stack.custom.d.ts
@@ -1,41 +1,0 @@
-export interface InterfaceStackAttributes {
-  appearance:
-    | 'vertical'
-    | 'horizontal'
-    | 'vertical--centered'
-    | 'horizontal--centered'
-    | 'horizontal--end';
-  collapseHorizontalBelow:
-    | 'micro'
-    | 'small'
-    | 'medium'
-    | 'large'
-    | 'wide'
-    | 'ultra'
-    | 'ultra--max-content';
-  spacing:
-    | 'fixed-1x'
-    | 'fixed-2x'
-    | 'fixed-3x'
-    | 'fixed-4x'
-    | 'fixed-5x'
-    | 'fixed-6x'
-    | 'fixed-8x'
-    | 'fixed-10x'
-    | 'fixed-12x'
-    | 'fixed-14x'
-    | 'fixed-16x'
-    | 'fixed-18x'
-    | 'fixed-20x'
-    | 'fixed-24x'
-    | 'fixed-30x'
-    | 'responsive-xxxs'
-    | 'responsive-xxs'
-    | 'responsive-xs'
-    | 'responsive-s'
-    | 'responsive-m'
-    | 'responsive-l'
-    | 'responsive-xl'
-    | 'responsive-xxl';
-  tag: 'div' | 'ul' | 'ol';
-}
