--- src/components/sbb-stack/sbb-stack.e2e.ts
@@ -1,13 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-
-describe('sbb-stack', () => {
-  let element, page;
-
-  it('renders', async () => {
-    page = await newE2EPage();
-    await page.setContent('<sbb-stack></sbb-stack>');
-
-    element = await page.find('sbb-stack');
-    expect(element).toHaveClass('hydrated');
-  });
-});
