+++ src/global/styles/_index.scss
@@ -7,7 +7,7 @@
 @forward './mixins/badge';
 @forward './mixins/buttons';
 @forward './mixins/card';
-@forward './mixins/grid_layout';
+@forward './mixins/grid-layout';
 @forward './mixins/helpers';
 @forward './mixins/link';
 @forward './mixins/lists';
