+++ src/global/styles/mixins/grid-layout.scss
@@ -1,0 +1,23 @@
+@use '../core/mediaqueries';
+
+@mixin grid-lean {
+  --sbb-grid-gap: var(--sbb-grid-base-gutter-responsive);
+  --sbb-grid-column-amount: var(--sbb-grid-base-columns);
+
+  display: grid;
+  gap: var(--sbb-grid-gap);
+  grid-template-columns: repeat(var(--sbb-grid-base-columns), 1fr);
+}
+
+@mixin grid {
+  @include grid-lean;
+
+  --sbb-grid-offset: var(--sbb-layout-base-offset-responsive);
+  --sbb-grid-max-width: var(--sbb-layout-base-page-max-width);
+
+  padding-inline: var(--sbb-grid-offset);
+  margin-inline: auto;
+
+  // Subtract padding to get real max-width
+  max-width: calc(var(--sbb-grid-max-width) + 2 * var(--sbb-grid-offset));
+}
