--- src/global/styles/mixins/grid_layout.scss
@@ -1,19 +1,0 @@
-@use '../core/mediaqueries';
-
-@mixin grid-base() {
-  display: grid;
-  grid-template-columns: repeat(var(--sbb-grid-base-columns), 1fr);
-  column-gap: var(--sbb-grid-base-gutter-responsive);
-  row-gap: var(--sbb-spacing-fixed-3x);
-}
-
-@mixin grid-base-top-products() {
-  display: grid;
-  column-gap: var(--sbb-grid-base-gutter-responsive);
-  row-gap: var(--sbb-spacing-fixed-3x);
-
-  @include mediaqueries.mq($from: medium) {
-    grid-template-columns: repeat(var(--sbb-grid-base-columns), 1fr);
-    max-width: 100%;
-  }
-}
