+++ src/storybook/styles/grid/readme.md
@@ -1,0 +1,16 @@
+# Grid
+
+Grid documentation.
+
+| css class       | sass mixin  |
+| --------------- | ----------- |
+| `sbb-grid`      | `grid`      |
+| `sbb-grid-lean` | `grid-lean` |
+
+### Usage
+
+```html
+<div class="sbb-grid">
+  <div class="item-on-grid"></div>
+</div>
+```
